# apps/pc-companion/src-tauri/resources/chatgpt-attach/openapi-cloud-control.yaml
openapi: 3.1.0
info:
  title: Cyclone Cloud Control API
  version: 0.2.0
  description: >-
    ChatGPT Actions front door for VMOS phones. Mutations execute through
    Cyclone Mobile PhoneToolExecutor. The short-lived SESSION_TOKEN from the
    latest Cyclone handoff is sent in X-Cyclone-Session-Token on every protected action.
servers:
  - url: https://CONTROL_API_HOST_PLACEHOLDER
    description: Cyclone One binds this to the current public CONTROL_API before you paste the schema.
components:
  parameters:
    CycloneSessionToken:
      name: X-Cyclone-Session-Token
      in: header
      required: true
      description: >-
        Short-lived SESSION_TOKEN from the latest CYCLONE_VMOS_ATTACH_v1 block.
        This is not an SSH Connect Key, VMOS AccessKey, or Cyclone One gateway token.
      schema:
        type: string
        minLength: 16
  schemas:
...
    ObserveResult:
      type: object
      properties:
        deviceId: { type: string }
        sessionId: { type: string }
        screenshotUrl: { type: string, format: uri }
        summary: { type: string }
        foregroundPackage: { type: string }
        observationId: { type: string }
        mutationEngine: { type: string }
paths:
...
  /v1/devices/{deviceId}/observe:
    post:
      operationId: observe
      summary: Capture a screenshot and compact UI summary through Cyclone
      parameters:
        - $ref: '#/components/parameters/CycloneSessionToken'
        - name: deviceId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId]
              properties:
                sessionId: { type: string }
                includeUiSummary: { type: boolean, default: true }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ObserveResult' }
...